Bring in latest slimfly synth from fitfly repo
This is a psuedo merge of the latest version of the slimfly
synthetic workload generator from Noah's fitfly specific codes
fork.

Most changes are "accept fitfly repo version" rule but a couple
changes were made to reflect changes made to the original version
that weren't also made to the forked version at the time.