Bring in updated slimfly model from fitfly fork
This commit brings in changes that were made to the slimfly
model by Noah since his fork to add support of the fitfly
variant. This introduces dual-rail/dual-plane behavior. It
should not interfere with single-rail/plane behavior of
the model if configured that way.

Almost all changes were incremental with only a few changes from
mainline that were brought in - specifically vis api and damaris
support changes.