mpi-replay: Add compute_time_speedup parameter
This commit adds functionality to accept a compute_time_speedup="X"
parameter in a model's configuration file to accelerate the delay
applied when accounting for compute_time in model-net-mpi-replay.c.

X here is a double where 2.0 represents a 2x reduction in delay
resulting from compute_time simulation. e.g. a setting of "2.0" means
that all increments of compute_time are reduced by a factor of 2.