fixup: compute_time_speedup event timestamp bug
The feature addition of the compute_time_speedup factor did affect
the compute time correctly. It, however, didn't affect the scheduling
of the new event after the compute time. This resulted in the delay
savings from the compute time speedup being applied to the comm. time
instead. This has now been resolved.