Slim Fly updates - Multi Rail Slim Fly
* Bring in latest slimfly synth from fitfly repo

This is a merge of the latest version of the slimfly
synthetic workload generator from Noah's fitfly specific codes
fork.

Most changes are "accept fitfly repo version" rule but a couple
changes were made to reflect changes made to the original version
that weren't also made to the forked version at the time.

* Bring in updated slimfly model from fitfly fork

This commit brings in changes that were made to the slimfly
model by Noah since his fork to add support of the fitfly
variant. This introduces dual-rail/dual-plane behavior. It
should not interfere with single-rail/plane behavior of
the model if configured that way.

Almost all changes were incremental with only a few changes from
mainline that were brought in - specifically vis api and damaris
support changes.

* Fix slimfly tests

When converting slimfly version to support fitfly, the names of the
workload LPs changed to nw-lp.

Additionally, changing the default sf_type to be 0 (slimfly) instead
of 1 (fitfly)

* Bring in codes changes for multi-injection

This commit brings in changes from Noah's fitfly fork. The changes
are all in the realm of allowing for multiple injection queues
in model terminals.

* Add model-net-synthetic-all to build

* fixes - comment out single proc calloc for heat map

* Include fix for injection throttling

* Undo last commit that breaks optimistic

The last commit that fixes the injection rate for dual plane
breaks optimisitc as it's a hack and not really valid. Rolling
it back while I try and find a legal way around it.

* dual rail rc fixes, malloc->calloc, inits

* small changes to slimfly/fitfly

* Slimfly Router Model Net Events Adaptation

* Slimfly/Fitfly RC Fixes - optimistic works

Prior to this commit, multi-rail slimfly was not working in optimistic
mode. Since it worked in sequential and conservative the issue was
surely in an RC. This bug in particular was caused by popping/pushing
from the incorrect queue in RC compared to FC.

Still not deterministic yet.

* Slimfly RC Fixes

* Fitfly event ties fix

* Update Slimfly Header for previous RC fixes

* Update tests for new slimfly model interface

* Slimfly update synthetic workload generators for fitfly unif random

* mpi-replay: add synthetic traffic print toggle

* DFDally/DFP add 3k interconnects

* Fitfly: arbitrary number of rails complete

* Fitfly: multirail adaptive routing complete

* model-net-synthetic-all generator timing change

* add bandwidth aggregation

* cn delay fix for multi rail

* synthetic-all: end time calculated based on last message received

* Fitfly fixes, add slow all to all traffic pattern, reverting rail path select

* slimfly/all workload: function prototype to make XL compiler happy

* Delete model-net-synthetic-all.c

* Update Makefile.subdir

* Update model-net-mpi-replay.c

* Remove old Fit Fly based comment from model-net-lp