DFDally: Add legacy prog adaptive routing
progress toward dfdally legacy prog-adaptive rewrite

DFDally: Legacy Progressive Adaptive Routing and fixes

This commit contains a few various code style fixes but most importantly
it includes the legacy progressive adaptive algorithm as a configurable
option for use. To use, specify routing="prog-adaptive-legacy"; in the
supplied configuration file.

Support for this algorithm should be considered ended. It is challenging
to maintain as the rationale behind routing decisions is not documented.
This means that it is difficult to tell the difference between
unintended and intended behavior in the code. We still wanted to allow
for this routing algorithm to be used if necessary but shouldn't be used
without knowing the potential risks of using it.

DFDally: fix uninitialized intermediate router in msg

DFDally: better adaptive threshold for config

DFDally: Cleanup