Dragonfly Models: Credit Delay Flexibility
Addresses issue #190. Previously, all dragonfly models used a static
value for the credit delay. It was calculated based on the local
bandwidth regardless of what type of link the credit was being
sent along. This was fine in networks where all links had the
same bandwidth but dragonfly models do not necessarily follow this
rule. This commmit adds features that allow for credit
delays to be set based on the type of link. They are also
configurable.

By default, the behavior remains the same as before. If the
configuration file supplied contains no credit delay specifications
then all credit delays will be based on the local bandwidth and the
credit size (default 8).

To make credit delays be set based on their respective bandwidths
(and credit size), then simply add `auto_credit_delay="1";` to
your config file.

If you'd like to set a single general credit delay for all, then
use `credit_delay="<float>";`. This is incompatible with the auto
credit delay functionality.

If you'd like to configure delays manually, then you can specify
`*_credit_delay="<float>";` for `cn`, `local`, and `global` in
place of the asterisk. If you set one of them but not others,
then the others will be auto calculated based on the bandwidth
and credit size.

The credit size is now also configurable via `credit_size="<int>";`
in bytes.

This commit also fixes the Dragonfly Plus/Dally print configuration
method to use fprintf so that the _IO_FILE* can be passed as a
parameter. This prevents issues where fprintf and printf were
overlapping during the initial parameter printout.