DFP Synthetic: Fix Single Job Segfault
The additions to make this synthetic workload generator allow multiple
synthetic workloads were done in a way that made it break for just a
standard synthetic run.

This was because the jobmap was not initialized for the case when there
was only a single job and so when that jobmap is polled to find what
job ID an MPI rank belongs to, it segfaults.

The solution to this is to initialize the jobmap_ctx value with
jobmap_ctx = codes_jobmap_configure(CODES_JOBMAP_IDENTITY, &jobmap_p);
where the IDENTITY jobmapper provides the funcionality so that if
the job ID lookup is called when there's only a single job, it doesn't
segfault.