Dragonfly Plus: Undo conditional output change
Commit a4780aed0 changed the behavior of the LPIO output files
generated by the Dragonfly Plus model so that if there was no
traffic on a port, then stats for that port would not be printed.

While this might make for a cleaner output file, it makes the
length of said file dependent on the simulation that it ran.
Comparing two output files visually by a human between two
different types of simulations becomes impossible.

It is also limiting in that it becomes more difficult to immediately
tell if there's a problem. For example, not long ago I noticed that
there was a port that wasn't getting any traffic due to a problem
in the routing implementation. This was easily visible because
there was a line with 0 traffic. If that line hadn't been printed
at all, it might not have been as easy to see had I not been looking
for it.

If this change was made to make creating graphs from non-negligible
data easier, then that is something that should be implemented as a
data pre-processing step and not hardcoded behavior in a generalized
simulation model.