DFP: Connection scoring now allows terminal conn
During a code review a long time ago when this change was initially
made, I suggested the TODO here. That was a bad on my part, it could
have just been implemented as such. Scoring a terminal connection
doesn't really mean much but maybe it could. At this point I see
no reason to not allow it.