DFP: Change way dfp_src_terminal_id is init
This commit changes the way that the recently introduced
dfp_src_terminal_id value encoded into messages is initialized.

Previously it was set based on codes_mapping_get_lp_relative_id()
relative to the server that generated the packet. This is
coincidentally the same as the terminal ID if there is only a
single server per terminal but this is not always the case.

It is now set during packet_generate() which is the function
called by the terminal that is generating the packet and so
we can cut out the call to codes_mapping_get_lp_relative_id()
entirely.