DFP Synthetic: Add note on major changes
I added a note here. There are major changes to this file, enough
that it probably warrants being a new file. I intend on unifying
synthetic workload generators going forward so this is more of a
note for anyone who wanted to have access to the old behavior
and a reminder to myself what the old version was when I'm making
the new workloads.