DFP: fix best connection randomizer
There was a previous fix that, if there were multiple 'best'
connections, it would try to pick randomly from them instead of just
picking the first one. This is a good idea but the implementation didn't
clear the vector of options when a better one was found.

As a result the vector that is randomly picked from will contain all
previous "best" options before a better one was found.

This fixes that.