Hotfix: Replace GNU _IO_FILE with FILE
Hotfix to replace the usage of _IO_FILE with FILE. This is what
happens when one uses some GNU based documentation exampled to modify
for the sake of time. _IO_FILE may not work with some compilers
like XLC, no reason to not just use FILE - like it should have been
before.