Rename Connection Manager File to Network Manager
The connection manager file has been renamed to network manager
as this file holds both the larger "NetworkManager" class as
well as the contained "ConnectionManager" class.

The rename makes it more obvious that this file has things that
are useful for the organization of a network and thus has
broader use than just managing connections on a router.