DFDally Synthetic: Random Permutation Behavioral Fix
This commit fixes the random permutation behavior to be in line with what
is expected in a random permutation traffic pattern. The previous implementation
did not actually do any random permutation as there was no counter to determine
how much data had been transmitted to the destination. Consequently there
was no conditional to check for said counter to switch to a new destination.
Each workload rank picked their own random destination and sent all of their
data to it.

The new implementation adds a command line argument --rperm_threshold= which
allows the user to set how much data is transmitted to a specific rank before
a new rperm destination is chosen by the sending workload rank.

This commit maintains the determinism added in the parent commit: 16b9a2d4f69e