Codes Online Workload: Fix deadlocking SWM_Sendrecv()
The previous implementation (#if else'd out) utilized a blocking
send followed by a blocking receive. If everyone in a comm executed
a sendrecv() with this strategy, then everyone will be stuck on
their blocking send and deadlock. This occurs with SWM LAMMPS if
rendezvous protocol is forced through a low eager threshold.

Results in unmatched isend irecv error as the simulation eventually
ends because no more events are issued due to deadlock.

OpenMPI documentation mentions the risk of this and any implementation
of the MPI_Sendrecv() pattern must take it into account. The old
implementation did not.

OpenMPI and MPICH both use some combination of nonblocking send/recv
and/or a waitall. This commit provides MPICH's isend+irecv+waitall
pattern.