Improve strategy to store latency of packets from terminal to terminal
Previously, the latency (delay) of the packet was assumed to be the
latency of the last chunk to arrive at the destination terminal. This is
wrong. We must store the time at which the first chunk is sent and the
time at which the last chunk is received. This change paves the way to
implement a strategy to feed a predictor with latencies in the order in
which the packets where sent (not delievered).