The state of the network simulation is now being *truly* frozen
This is buggy. With some models this will work, with others it will
definetely not. The problem resides on model-net's complexity. Knowing
when to trigger the "next event" event is dependent on the state of
model-net and its future messages already in the queue (the most
important of which is the workload's new event)