Simplifying zombie notification strategy
The previous strategy required to keep track of all packets on arrival.
This was done to prevent duplicated packets from being processed at the
destination terminal. A notification of arrival is sent once the packet
is processed, but this notification in not zero-offset, so the thought
was "when switching to surrogate, a source terminal might not get the
notification a notification (it's scheduled for a future point in time)
and thus the source terminal will assume that the event has not being
processed and has to be notified of being a zombie, which means that we
have to keep track of all packets processed at the destination to
prevent duplicates".

The reasoning was correct, but there is another workaround. We can look
at the future and see if the notification has been sent, yet not
processed by the source terminal! This reduces the complexity on the
network model side as well as reduces memory consumption and time :).