New strategy to feed predictor implemented
Previously, the predictor would only be given packet information (packet
id, latency, etc) in an increasing order on packet id. Now it's
different. The predictor is fed the packet information whenever is
available.

Sometimes, the first packet injected in the network takes far too long
to be delievered. In those cases, the predictor never receives any info
for long periods of time.