Forcing variable to not be optimized out
A variable was declared and defined under a scope that was disregarded
later on when the variable was used again.

This expression was never executed:

> oc_params.nprocs = num_traces_of_job[lid.job];

because `oc_params.nprocs` was never executed. The line:

> params = (char*)&oc_params;

after was copying the `oc_params` variable into a new variable to be
used later. `oc_params` is not guaranteed to exist after the scope
closes, and this would provoke a segfault.