Fix: replacing O(n) table lookup for O(1)
When calling the function `jobmap_list_to_local`, this would go through
the entire list of IDs until it finds a matching ID. This is O(n) in the
average and worst cases. For small networks, this won't take much time,
so it never flared up as an issue. When running larger network
simulations, at 8K nodes, there was a significant slowdown. This
function was found, after extensive profiling, to be the principal
culprit.

The fix is simple, make a table where looking for an ID is O(1). A
simple array does the trick. After running some experiments, there's a
significant speedup of 30% for a network of 8448 with a job using all
nodes. The job was uniform random and the simulation was run for 10ms
(virtual time).