Fixing compilation warning `incompatible-pointer-types`
This bug was introduced when building the network surrogate. To build
the surrogate, we need to track the input queue "size" (the input
message queue to the routers from the workloads).

If the network surrogate wouldn't live down in specific network models
(it has been implemented right now only on dragonfly-dally), it should
actually reside within the model-net layer, and thus, individual models
shouldn't need to track the state of the input queue.

Hopefully, we can move the network surrogate from dragonfly-dally into
model-net.